refactor: Command 빌더 패턴 적용
(cherry picked from commit 9f34f2b82a8e1541306b33849a3aba14f5f88d41)
