+++ backend/src/main/java/com/festago/admin/dto/artist/ArtistV1CreateRequest.java
@@ -13,10 +13,10 @@ public record ArtistV1CreateRequest(
 ) {
 
     public ArtistCreateCommand toCommand() {
-        return new ArtistCreateCommand(
-            name,
-            profileImageUrl,
-            backgroundImageUrl
-        );
+        return ArtistCreateCommand.builder()
+            .name(name)
+            .profileImageUrl(profileImageUrl)
+            .backgroundImageUrl(backgroundImageUrl)
+            .build();
     }
 }
