+++ backend/src/main/java/com/festago/admin/dto/artist/ArtistV1UpdateRequest.java
@@ -13,10 +13,10 @@ public record ArtistV1UpdateRequest(
 ) {
 
     public ArtistUpdateCommand toCommand() {
-        return new ArtistUpdateCommand(
-            name,
-            profileImageUrl,
-            backgroundImageUrl
-        );
+        return ArtistUpdateCommand.builder()
+            .name(name)
+            .profileImageUrl(profileImageUrl)
+            .backgroundImageUrl(backgroundImageUrl)
+            .build();
     }
 }
