+++ backend/src/main/java/com/festago/admin/dto/stage/StageV1UpdateRequest.java
@@ -18,10 +18,10 @@ public record StageV1UpdateRequest(
 ) {
 
     public StageUpdateCommand toCommand() {
-        return new StageUpdateCommand(
-            startTime,
-            ticketOpenTime,
-            artistIds
-        );
+        return StageUpdateCommand.builder()
+            .startTime(startTime)
+            .ticketOpenTime(ticketOpenTime)
+            .artistIds(artistIds)
+            .build();
     }
 }
