+++ backend/src/main/java/com/festago/artist/dto/command/ArtistCreateCommand.java
@@ -1,5 +1,8 @@
 package com.festago.artist.dto.command;
 
+import lombok.Builder;
+
+@Builder
 public record ArtistCreateCommand(
     String name,
     String profileImageUrl,
