+++ backend/src/main/java/com/festago/auth/dto/command/AdminSignupCommand.java
@@ -1,5 +1,8 @@
 package com.festago.auth.dto.command;
 
+import lombok.Builder;
+
+@Builder
 public record AdminSignupCommand(
     String username,
     String password
