+++ backend/src/main/java/com/festago/festival/dto/command/FestivalUpdateCommand.java
@@ -1,7 +1,9 @@
 package com.festago.festival.dto.command;
 
 import java.time.LocalDate;
+import lombok.Builder;
 
+@Builder
 public record FestivalUpdateCommand(
     String name,
     LocalDate startDate,
