+++ backend/src/main/java/com/festago/stage/dto/command/StageCreateCommand.java
@@ -2,7 +2,9 @@ package com.festago.stage.dto.command;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import lombok.Builder;
 
+@Builder
 public record StageCreateCommand(
     Long festivalId,
     LocalDateTime startTime,
