+++ backend/src/main/java/com/festago/stage/dto/command/StageUpdateCommand.java
@@ -2,7 +2,9 @@ package com.festago.stage.dto.command;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import lombok.Builder;
 
+@Builder
 public record StageUpdateCommand(
     LocalDateTime startTime,
     LocalDateTime ticketOpenTime,
